spotify-backend-service/pyproject.toml
[project]
name = "spotify-backend"
version = "0.1.0"
dependencies = [
  "grpcio",
  "grpcio-tools",
  "requests",
  "redis",
  "python-dotenv"
]
